Workout Get, Meal Get

+++ BuiltDifferentMobileApp/BuiltDifferentMobileApp/Models/Meal.cs
@@ -6,8 +6,8 @@ namespace BuiltDifferentMobileApp.Models {
     public class Meal {
         public int id { get; set; }
         public string mealName { get; set; }
-        public int calories { get; set; }
-        public int protein { get; set; }
+        public double calories { get; set; }
+        public double protein { get; set; }
         public double carbs { get; set; }
         public double fat { get; set; }
         public string recipe { get; set; }
